+++ packages/frontend/src/components/PrivateRoute.tsx
@@ -1,8 +1,8 @@
-import React, { ComponentProps, FC, useEffect } from 'react';
+import { ComponentProps, FC, useEffect } from 'react';
 import { Redirect, Route } from 'react-router-dom';
 import { useEmailStatus } from '../hooks/useEmailVerification';
 import { useApp } from '../providers/AppProvider/useApp';
-import { useAbilityContext } from './common/Authorization';
+import { useAbilityContext } from './common/Authorization/useAbilityContext';
 import PageSpinner from './PageSpinner';
 
 const PrivateRoute: FC<ComponentProps<typeof Route>> = ({
