+++ packages/frontend/src/components/UserSettings/AppearancePanel/index.tsx
@@ -14,7 +14,8 @@ import { FC, useCallback, useEffect } from 'react';
 import { useOrganization } from '../../../hooks/organization/useOrganization';
 import { useOrganizationUpdateMutation } from '../../../hooks/organization/useOrganizationUpdateMutation';
 import { isHexCodeColor } from '../../../utils/colorUtils';
-import { Can, useAbilityContext } from '../../common/Authorization';
+import { Can } from '../../common/Authorization';
+import { useAbilityContext } from '../../common/Authorization/useAbilityContext';
 
 const getColorFormFields = (colors: string[]) =>
     colors.reduce(
