+++ packages/frontend/src/components/common/Authorization/index.tsx
@@ -1,16 +1,6 @@
 import { Ability } from '@casl/ability';
 import { createContextualCan } from '@casl/react';
-import { createContext, useContext } from 'react';
+import { createContext } from 'react';
 
 export const AbilityContext = createContext(new Ability());
 export const Can = createContextualCan(AbilityContext.Consumer);
-
-export function useAbilityContext() {
-    const context = useContext(AbilityContext);
-    if (context === undefined) {
-        throw new Error(
-            'useAbilityContext must be used within a AbilityContext.Provider',
-        );
-    }
-    return context;
-}
